requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
blinker==1.4
certifi==2022.12.7
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
flask-bootstrap==3.3.7.1
flask-login==0.5.0
flask-mail==0.9.1
flask-moment==1.0.2
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
flask==2.0.1
greenlet>=3.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
jinja2==3.0.1
markupsafe==2.0.1
python-dotenv==0.19.0
requests==2.26.0
setuptools==67.6.1; python_version >= '3.7'
sqlalchemy==1.4.23
stripe==5.0.0
urllib3==1.26.7
visitor==0.1.3
werkzeug==2.0.1
wtforms==2.3.3
